+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Versions
 
+## Version 4.1.0
+
+- `getAccountingReport`: rework invoices and payments to show circular payments
+- `getAccountingReport`: change the formatting of invoice notes to be more terse
+
 ## Version 4.0.0
 
 - `getAccountingReport`: A request func (like @alexbosworth/request) is required